# Optimize rendering of long torrent lists