@dareiff 2023-10-24
I love it in theory, but currently see a breaking change to layout for each torrent cell:

<img width="505" alt="Screenshot 2023-10-24 at 9 19 38 AM" src="https://github.com/transmission/transmission/assets/435896/0ef0be4e-4f3d-45d2-988d-2a1fdc59938d">