@anthonyryan1 2023-10-24
That would appear the rows are a different size on your device than mine, and that does make this worse.

Which browser are you using there?

I'll see if there's a way I can express the height of the row that's consistent across all those browsers, so far I've tested Firefox and Chrome, should probably spin up a macOS VM and test Safari.