@anthonyryan1 2023-11-07
I'm scrapping this project for now.

There's enough open pull requests modifying the height of the rows, and this would make their work harder. If the height of the rows was relatively constant in all browsers and across many years of releases this is a good idea, but there's features in that screenshot I neglected to test (like the tags in titles).

I might revisit this someday in the future.